elimiinating the uploadProfilePic component moving to signup modal
