--- client/src/components/ProfilePicUpload.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactS3 from 'react-s3';
-import { uploadFile } from 'react-s3';
-
-const config = {
-    bucketName: process.env.REACT_APP_BUCKETNAME,
-    region: process.env.REACT_APP_REGION,
-    accessKeyId: process.env.REACT_APP_ACCESSKEYID,
-    secretAccessKey: process.env.REACT_APP_SECRETACCESSKEY
-}
